requirements.txt
PyGithub==2.1.1
click==8.1.7
pyyaml==6.0.1
tabulate==0.9.0
